pl_bot_project/requirements.txt
Flask==2.0.1
torch==1.9.0
transformers==4.9.0
faiss-cpu==1.7.0
langchain==0.0.10
pandas==1.3.3
scikit-learn==0.24.2
numpy==1.21.0
pyspellchecker==0.6.2
